Refactor, guard against re-runs, add diags header, tidy up
